Adjust toString(..) function of cell descriptions according to new neighbour merge performed data type.

+++ ExaHyPE/exahype/records/ADERDGCellDescription.cpp
@@ -1484,9 +1484,9 @@ switch (mode) {
       out << ",";
       out << "neighbourMergePerformed:[";
    for (int i = 0; i < DIMENSIONS_TIMES_TWO-1; i++) {
-      out << getNeighbourMergePerformed(i) << ",";
+      out << static_cast<int>(getNeighbourMergePerformed(i)) << ",";
    }
-   out << getNeighbourMergePerformed(DIMENSIONS_TIMES_TWO-1) << "]";
+   out << static_cast<int>(getNeighbourMergePerformed(DIMENSIONS_TIMES_TWO-1)) << "]";
       out << ",";
       out << "hasCompletedTimeStep:" << getHasCompletedTimeStep();
       out << ",";
